+++ znet/server.go
@@ -130,7 +130,7 @@ func (s *Server) Serve(c *gin.Context) {
 }
 
 //AddRouter 路由功能：给当前服务注册一个路由业务方法，供客户端链接处理使用
-func (s *Server) AddRouter(msgID uint32, router ziface.IRouter) {
+func (s *Server) AddRouter(msgID uint16, router ziface.IRouter) {
 	s.msgHandler.AddRouter(msgID, router)
 }
